+++ changelog.md
@@ -1,6 +1,7 @@
 v1.6.0
 ------
 * Fixed a mod compatibility bug in ore detection logic
+* Fixed an ore-matching issue with the Ore Diviner
 
 v1.5.0
 ------